Vistas cuidador y navegador completado

+++ recuerdame/listadoRecuerdos.php
@@ -98,8 +98,10 @@
                             <td><?php echo ($row["nombreEtiqueta"]) ?></td>
                             <td class="tableActions">
                                 <a href="verDatosRecuerdo.php?idRecuerdo=<?php echo ($row['idRecuerdo']) ?>"><i class="fa-solid fa-eye text-black tableIcon"></i></a>
-                                <a href="modificarDatosRecuerdo.php?idRecuerdo=<?php echo ($row['idRecuerdo']) ?>"><i class="fa-solid fa-pencil text-primary tableIcon"></i></a>
-                                <a href="gestor.php?accion=eliminarRecuerdo&idRecuerdo=<?php echo ($row['idRecuerdo']) ?>"><i class="fa-solid fa-trash-can text-danger tableIcon"></i></a>
+                                    <?php if(Session::esTerapeuta()) { ?> 
+                                        <a href="modificarDatosRecuerdo.php?idRecuerdo=<?php echo ($row['idRecuerdo']) ?>"><i class="fa-solid fa-pencil text-primary tableIcon"></i></a>
+                                        <a href="gestor.php?accion=eliminarRecuerdo&idRecuerdo=<?php echo ($row['idRecuerdo']) ?>"><i class="fa-solid fa-trash-can text-danger tableIcon"></i></a>
+                                    <?php } ?>
                             </td>
                         </tr>
                     <?php
